refactor: Phase 3 페이지 전용 컴포넌트 Colocation
- myInfo 컴포넌트 → pages/my-info/_components/ 이동
- myProject 컴포넌트 → pages/my-project/_components/ 이동
- report 컴포넌트 → pages/report/_components/ 이동
- profileDetail 컴포넌트 → pages/developer-detail/_components/ 이동
- SearchTabs → pages/search/_components/ 이동
- MyProjectTabs → pages/my-project/_components/ 이동
- DatePickerPopover, PositionBasedTechStackDropdown → pages/project-create/_components/ 이동
- recommend 드롭다운은 여러 페이지에서 공유되어 shared/에 유지
- 모든 import 경로 업데이트 완료, 빌드 성공
