+++ src/pages/developer-detail/_components/DomainBadges.tsx
