+++ src/pages/developer-detail/_components/ImagePreview.tsx
