+++ src/pages/developer-detail/_components/ReportCardSmall.tsx
