+++ src/pages/developer-detail/_components/TechStackChips.tsx
