+++ src/pages/developer-detail/developer-suggest-page.tsx
@@ -8,8 +8,8 @@ import BackIcon from '@assets/icons/back.svg?react';
 import { useAuth } from '@clerk/clerk-react';
 import MainProjectCard from '@components/common/MainProjectCard';
 import SelectDropdown from '@components/common/SelectDropdown';
-import ImagePreview from '@components/profileDetail/ImagePreview';
-import RoleChips from '@components/profileDetail/RoleChips';
+import ImagePreview from './_components/ImagePreview';
+import RoleChips from './_components/RoleChips';
 import { DOMAIN_REVERSE_MAP } from '@constants/domain';
 import { cn } from '@libs/cn';
 import { useQuery } from '@tanstack/react-query';
