+++ src/pages/my-info/my-info-profile-edit.tsx
@@ -11,9 +11,9 @@ import CheckboxCheckedIcon from '@assets/icons/checkbox-checked.svg?react';
 import CheckboxUncheckedIcon from '@assets/icons/checkbox-unchecked.svg?react';
 import PlusNolineIcon from '@assets/icons/plus-noline.svg?react';
 import { useAuth } from '@clerk/clerk-react';
-import MyInfoInput from '@components/myInfo/MyInfoInput';
-import StackChips from '@components/myInfo/StackChips';
-import ImagePreview from '@components/profileDetail/ImagePreview';
+import MyInfoInput from './_components/MyInfoInput';
+import StackChips from './_components/StackChips';
+import ImagePreview from '../developer-detail/_components/ImagePreview';
 import PositionTechStackDropdown from '@components/recommend/PositionTechStackDropdown';
 import { DOMAIN_MAP, DOMAIN_OPTIONS, DOMAIN_REVERSE_MAP } from '@constants/domain';
 import { getTechstackIdsByKeys } from '@constants/signup-mapping';
