+++ src/pages/my-info/my-info-setting.tsx
@@ -3,8 +3,8 @@ import { myInfoQueries, updateMyProfile } from '@apis/myInfo/myInfo-queries';
 import GithubIcon from '@assets/icons/github.svg?react';
 import GoogleIcon from '@assets/icons/google.svg?react';
 import { useUser } from '@clerk/clerk-react';
-import Switch from '@components/myInfo/Switch';
-import TabMenu from '@components/myInfo/TabMenu';
+import Switch from './_components/Switch';
+import TabMenu from './_components/TabMenu';
 import { cn } from '@libs/cn';
 import { useAuthStore } from '@store/auth';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
