+++ src/pages/recommend/recommend-page.tsx
@@ -1,6 +1,6 @@
 import { useAuth } from '@clerk/clerk-react';
 import LoginRequiredCard from '@components/common/LoginRequiredCard';
-import SearchTabs from '@components/tab/SearchTabs';
+import SearchTabs from '../search/_components/SearchTabs';
 import { Outlet, useLocation } from 'react-router-dom';
 
 const RecommendPage = () => {
