+++ src/pages/report/_components/Blank.tsx
