+++ src/pages/report/_components/ContentBox.tsx
