+++ src/pages/report/report-create-page.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useInfiniteQuery } from '@tanstack/react-query';
 
-import CheckBox from '@components/report/CheckBox';
-import GithubRepoListSkeleton from '@components/report/GithubRepoListSkeleton';
+import CheckBox from './_components/CheckBox';
+import GithubRepoListSkeleton from './_components/GithubRepoListSkeleton';
 import { cn } from '@libs/cn';
 import { useThemeStore } from '@store/theme';
 import { myInfoQueries } from '@apis/myInfo/myInfo-queries';
