+++ src/pages/report/report-result-page.tsx
@@ -1,4 +1,4 @@
-import ReportCard from '@components/report/ReportCard';
+import ReportCard from './_components/ReportCard';
 import { cn } from '@libs/cn';
 import { useThemeStore } from '@store/theme';
 import { useLocation, useNavigate } from 'react-router-dom';
